refactor: New utils file for ListObjectsV2.
