+++ src/s3_frontend/mod.rs
@@ -2,5 +2,5 @@ pub mod auth;
 //pub mod data_handler;
 pub mod s3server;
 pub mod s3service;
-//pub mod utils;
 pub mod impersonating_client;
+pub mod utils;
