+++ src/s3_frontend/utils/utils.rs
@@ -187,3 +187,4 @@ mod tests {
         );
     }
 }
+
